requirements.txt
Django~=4.2.5
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.1.0
django-cors-headers~=3.13.0
django-location-field~=2.7.3
geopy